// mifos-web-next/apps/web/src/lib/fineract/transfer-beneficiary-clients.types.ts
/**
 * Copyright since 2026 Mifos Initiative
 *
 * This Source Code Form is subject to the terms of the Mozilla Public
 * License, v. 2.0. If a copy of the MPL was not distributed with this
 * file, You can obtain one at http://mozilla.org/MPL/2.0/.
 */

export type TransferBeneficiaryClientSummary = {
  id: number;
  displayName: string;
  accountNo?: string;
  officeName?: string;
};

export type TransferBeneficiaryClientResolved = TransferBeneficiaryClientSummary & {
  officeId: number;
};